Refactored $lib/quiz
- Removed $lib/constants/quiz
- Isolated quiz data in separate file
- Created quiz responses storage service
- Isolated types in separate file
- Updated changed imports

Note: All quiz-related code shall be exported from the index.ts file e.g. from $lib/quiz.
